chore: remove support for toggle-able embedded bot
