+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@snf/access-qa-bot",
-  "version": "1.0.0-beta.1",
+  "version": "1.1.0-beta.1",
   "private": false,
   "homepage": ".",
   "description": "ACCESS Q&A Bot React Component and Web Component",
