+++ web-component-demo.html
@@ -37,6 +37,7 @@
             welcome="Welcome to the Q&A Bot!"
             prompt="Ask me anything about ACCESS..."
             is-logged-in
+            default-open
             api-key="demo-key">
         </access-qa-bot>
     </div>
@@ -71,7 +72,6 @@
                     prompt: "Ask a question about ACCESS...",
                     isLoggedIn: true,
                     embedded: true,
-                    defaultOpen: false,
                     apiKey: 'demo-key'
                 });
             }
